fix #1655 Jaws reads "not checked unavailable" when the checkbox is checked (and disabled)
*****
NOT WORKING YET!
TODO: make the Jaws test succeed
*****

This commit contains a test for the following bug in IE 11 / Jaws 16:
in some cases a disabled checked checkbox is read as "not checked unavailable"
instead of "checked unavailable". It seems to happen when aria attributes
(such as aria-disabled or aria-labelledby) are used on the input.

PTR 11793963