feat #1759 Adding support for beans pre-compilation
This commit adds support for beans pre-compilation.
The new aria.util.BeanExtractor.extract method allows to retrieve from the
JsonValidator a compiled beans package as a javascript file to be used
instead of the original file containing Aria.beanDefinitions. The compiled
file is usually bigger than the original file, but is quicker to execute.

close #1759