fix ATDebug flag introduced in #1759 and error messages in BeanExtractor
This commit removes the ATDebug flag introduced in #1759 because it caused
all templates to be compiled in debug mode (with line numbers).
This commit also adds new checks in BeanExtractor to be sure that the
corresponding configuration in the JsonValidator is correct for the
extraction with the given options to be successful.